requirements.txt
# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.25.0
transformers>=4.36.0

# Data and caching (compatible versions)
datasets>=2.18.0
huggingface_hub>=0.20.0
safetensors>=0.4.0

# Training
hydra-core>=1.3.0
omegaconf>=2.3.0
accelerate>=0.25.0

# Logging
wandb>=0.16.0
tensorboard>=2.15.0

# Video processing
opencv-python>=4.9.0
av>=11.0.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
pillow>=10.0.0

# Additional missing dependencies
pyarrow>=15.0.0
fsspec>=2023.10.0
